Extract stacking test cases and utils
